♻️ Refactor functions for editing into OgsmList
